+++ LibClassicSwingTimerAPI.lua
@@ -160,7 +160,7 @@ function Unit:SwingEnd(hand)
 	end
 	if self.class == "DRUID" and self.skipNextAttackSpeedUpdate then
 		self.skipNextAttackSpeedUpdate = nil
-		lib:UNIT_ATTACK_SPEED(self.GUID)
+		lib:UNIT_ATTACK_SPEED('UNIT_ATTACK_SPEED', self.GUID)
 	end
 	self.callbacks:Fire("UNIT_SWING_TIMER_STOP", self.id, hand)
 	if (self.casting or self.channeling) and self.isAttacking and hand ~= "ranged" then
@@ -385,7 +385,7 @@ function lib:COMBAT_LOG_EVENT_UNFILTERED(_, ts, subEvent, _, sourceGUID, _, _, _
 	end
 end
 
-function lib:UNIT_ATTACK_SPEED(unitGUID)
+function lib:UNIT_ATTACK_SPEED(_, unitGUID)
 	local unit = lib:getUnit(unitGUID)
 	if not unit then
 		return
@@ -496,6 +496,8 @@ function lib:UNIT_SPELLCAST_SUCCEEDED(_, unitType, _, spell)
 		if isRetail then		
 			unit:SwingStart("mainhand", now, true)
 		else
+			-- Do not skip the attack speed update if we reset the timer
+			unit.skipNextAttackSpeedUpdate = nil
 			unit:SwingStart("mainhand", now, true)
 			unit:SwingStart("offhand", now, true)
 		end